Refactor to node_uuid so it doesnt conflict with built-in uuid
