+++ pyblish_magenta/utils/maya/node_uuid.py
